+++ lib/api/routes/fetch.ts
@@ -19,8 +19,6 @@ const routeFetch = (zamza: Zamza) => {
                 "/api/fetch/:topic/range/earliest/:count",
                 "/api/fetch/:topic/paginate/stoe/:skipToIndex/:limit",
                 "/api/fetch/:topic/paginate/etos/:skipToIndex/:limit",
-                "/api/fetch/:topic/query/find",
-                "/api/fetch/:topic/query/count",
             ],
         });
     });
@@ -187,50 +185,6 @@ const routeFetch = (zamza: Zamza) => {
         try {
             res.status(200).json(await keyIndexModel.paginateThroughTopic(topic,
                 skipToIndex ? skipToIndex : null, parseInt(limit, undefined), -1));
-        } catch (error) {
-            res.status(500).json({
-                error: "An error occured " + error.message,
-            });
-        }
-    });
-
-    router.post("/:topic/query/find", async (req, res) => {
-
-        const {topic} = req.params;
-        const {query, limit = 512, skipToIndex = null, order = -1} = req.body;
-
-        if (!res.locals.access.topicAccessAllowedForRequest(req, topic)) {
-            res.status(403).json({
-                error: "Access not allowed, for this topic",
-            });
-            return;
-        }
-
-        try {
-            res.status(200).json(await keyIndexModel.findForQuery(topic, query, limit, skipToIndex, order));
-        } catch (error) {
-            res.status(500).json({
-                error: "An error occured " + error.message,
-            });
-        }
-    });
-
-    router.post("/:topic/query/count", async (req, res) => {
-
-        const {topic} = req.params;
-        const {query, limit = 512, skipToIndex = null, order = -1} = req.body;
-
-        if (!res.locals.access.topicAccessAllowedForRequest(req, topic)) {
-            res.status(403).json({
-                error: "Access not allowed, for this topic",
-            });
-            return;
-        }
-
-        try {
-            res.status(405).json({
-                error: "Not implemented",
-            });
         } catch (error) {
             res.status(500).json({
                 error: "An error occured " + error.message,