+++ lib/api/routes/index.ts
@@ -7,3 +7,4 @@ export * from "./manage";
 export * from "./hook";
 export * from "./replay";
 export * from "./state";
+export * from "./query";